config.fs: Specify firmware image files as system:system Explicitly specify firmware image files legacy /firmware/image/* as system:system and new /vendor/firmware_mnt/image/* as root:system
For normal boot, this is a no-op as /firmware|/vendor/firmware_mnt
is a mount point.  This change is only for development when
firmware is read-writable and allows changes via "adb push".  This
match the underlying FAT fs mount ownership of system:system or
root:system so adb push's chown will not error out.